+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # Changelog
+
+## 0.1.0
+
+* Add more validation, to valid string.
+* Define words as chars with length over 1.
+* Trim each line to avoid misleading spaces.